+++ src/components/index.ts
@@ -3,3 +3,4 @@ export * from './CropImage';
 export * from './DropZone';
 export * from './ImageWrapper';
 export * from './TableData';
+export * from './MenuBar';
